+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2.1.0 (2019-06-??)
+
+### Added
+
+- 
+
+### Bugs
+
+- `VideoConvertParams` throws `InvalidParamException` instead of InvalidArgumentException 
+
 ## 2.0.4 (2019-04-11)
 
 ### Bugs